Fix FC number not showing
- show FC number if duration is 0 but its not the first cycle
- move BallotStateBadge to own file
- minor copy
